% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsvar-package.R
\docType{package}
\name{bsvar-package}
\alias{bsvar-package}
\alias{bsvar}
\title{The 'bsvar' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.11. https://mc-stan.org
}
