% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{extract}
\alias{extract}
\title{Extract samples from a fitted model}
\usage{
extract(object, pars, apply_restriction = TRUE)
}
\arguments{
\item{object}{An instance of class \code{stanfit} as outputted by \code{\link{bsvar}}.}

\item{pars}{A character vector of parameter names.If not specified, all parameters
and other quantities are used.}

\item{apply_restriction}{Logical. Defaults to \code{TRUE} in which case the posterior
sample is restricted according to a binary vector given by \code{attributes(object)$restriction},
if exists (typically it does not), such that ones stand for posterior draws to be dropped from
the restricted sample. This can be used for imposing sign restrictions a posteriori.}
}
\description{
Wrapper for \code{\link[rstan]{extract}}, but with an additional argument
(\code{apply_restriction}).
}
\seealso{
\code{\link[rstan]{extract}}
}
