% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irf}
\alias{irf}
\title{\code{irf} computes and plots impulse response function for one shock.}
\usage{
irf(
  obj,
  horizon = 40,
  sub_N = NULL,
  shock = 1,
  variables = "all",
  cumulative = c(),
  shock_size = 1,
  fixed_impact = 0,
  transform = NULL,
  controls = list(),
  apply_restriction = TRUE,
  R_progress_bar = TRUE
)
}
\description{
\code{irf} computes and plots impulse response function for one shock.
}
