% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irf.R
\name{irfs}
\alias{irfs}
\title{\code{irfs} computes and plots impulse response functions for multiple shocks.}
\usage{
irfs(
  obj,
  horizon = 40,
  sub_N = NULL,
  shocks = "all",
  variables = "all",
  cumulative = c(),
  shock_sizes = 1,
  fixed_impacts = 0,
  transform = NULL,
  controls = list(),
  apply_restriction = TRUE,
  R_progress_bar = TRUE
)
}
\description{
\code{irfs} computes and plots impulse response functions for multiple shocks.
}
