% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{rhat}
\alias{rhat}
\title{\loadmathjax{}
\mjseqn{\hat{R}} convergence diagnostic for Markov Chains}
\usage{
rhat(object, pars = NULL)
}
\arguments{
\item{object}{An instance of class \code{stanfit} as outputted by \code{\link{bsvar}}.}

\item{pars}{A character vector of parameter names. Defaults to a predefined set of the
most important parameters.}
}
\description{
A convenience function that extracts \mjseqn{\hat{R}} convergence diagnostic values
from the summary of the \code{stanfit} object outputted by \code{\link{bsvar}}
and returns them as a named vector.
}
\seealso{
\code{\link[rstan]{Rhat}}
}
