% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock_decomp.R
\name{shock_decomp}
\alias{shock_decomp}
\title{\code{shock_decomp} computes ... for models
estimated with \code{bsvar}.}
\usage{
shock_decomp(obj, sub_N = NULL, apply_restriction = TRUE)
}
\arguments{
\item{obj}{An object of class `stanfit` as returned by `bsvar()`.}
}
\value{
...
}
\description{
\code{shock_decomp} computes ... for models
estimated with \code{bsvar}.
}
