% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{prnt}
\alias{prnt}
\title{Print a concise summary}
\usage{
prnt(
  x,
  pars = NULL,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  digits_summary = 2,
  include = TRUE,
  use_cache = FALSE,
  ...
)
}
\description{
Wrapper for \code{\link[rstan]{print}}, that prints by default a summary
of only some of the parameters from the \code{stanfit} object outputted
by \code{\link{bsvar}}. By default \code{use_cache = FALSE}, which makes \code{prnt} typically
much faster than \code{\link[rstan]{print}} if there is no cached data (right after estimation),
but much slower if there is cached data (if \code{\link[rstan]{print}} has already been called).
}
\seealso{
\code{\link[rstan]{print.stanfit}}
}
